+++ src/Exceptions/InvalidCartConfigurationException.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Contains the InvalidCartConfigurationException class.
+ *
+ * @copyright   Copyright (c) 2018 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2018-03-15
+ *
+ */
+
+namespace Vanilo\Cart\Exceptions;
+
+
+class InvalidCartConfigurationException extends \Exception
+{
+}