+++ src/Models/Cart.php
@@ -15,12 +15,14 @@ namespace Vanilo\Cart\Models;
 use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Collection;
-use InvalidArgumentException;
+use Vanilo\Cart\Exceptions\InvalidCartConfigurationException;
 use Vanilo\Contracts\Buyable;
 use Vanilo\Cart\Contracts\Cart as CartContract;
 
 class Cart extends Model implements CartContract
 {
+    const EXTRA_PRODUCT_MERGE_ATTRIBUTES_CONFIG_KEY = 'vanilo.cart.extra_product_attributes';
+
     protected $guarded = ['id'];
 
     /**
@@ -59,12 +61,13 @@ class Cart extends Model implements CartContract
             $item->quantity += $qty;
             $item->save();
         } else {
-            $item = $this->items()->create([
-                'product_type' => $product->morphTypeName(),
-                'product_id'   => $product->getId(),
-                'quantity'     => $qty,
-                'price'        => $product->getPrice()
-            ]);
+            $item = $this->items()->create(
+                array_merge(
+                    $this->getDefaultCartItemAttributes($product, $qty),
+                    $this->getExtraProductMergeAttributes($product),
+                    $params['attributes'] ?? []
+                )
+            );
         }
 
         $this->load('items');
@@ -140,5 +143,65 @@ class Cart extends Model implements CartContract
         }
 
         $this->user_id = $user;
+    }
+
+    /**
+     * Returns the default attributes of a Buyable for a cart item
+     *
+     * @param Buyable $product
+     * @param integer $qty
+     *
+     * @return array
+     */
+    private function getDefaultCartItemAttributes(Buyable $product, $qty)
+    {
+        return [
+            'product_type' => $product->morphTypeName(),
+            'product_id'   => $product->getId(),
+            'quantity'     => $qty,
+            'price'        => $product->getPrice()
+        ];
+    }
+
+
+    /**
+     * Returns the extra product merge attributes for cart_items based on the config
+     *
+     * @param Buyable $product
+     *
+     * @throws InvalidCartConfigurationException
+     *
+     * @return array
+     */
+    private function getExtraProductMergeAttributes(Buyable $product)
+    {
+        $result = [];
+        $cfg    = config(self::EXTRA_PRODUCT_MERGE_ATTRIBUTES_CONFIG_KEY, []);
+
+        if (!is_array($cfg)) {
+            throw new InvalidCartConfigurationException(
+                sprintf(
+                    'The value of `%s` configuration must be an array',
+                    self::EXTRA_PRODUCT_MERGE_ATTRIBUTES_CONFIG_KEY
+                )
+            );
+        }
+
+        foreach ($cfg as $attribute) {
+
+            if (!is_string($attribute)) {
+                throw new InvalidCartConfigurationException(
+                    sprintf(
+                        'The configuration `%s` can only contain an array of strings, `%s` given',
+                        self::EXTRA_PRODUCT_MERGE_ATTRIBUTES_CONFIG_KEY,
+                        gettype($attribute)
+                    )
+                );
+            }
+
+            $result[$attribute] = $product->{$attribute};
+        }
+
+        return $result;
     }
 }